Nicolas Regnault: - fix a bug with Hilbert space larger than 2^31 in AbstractPrecalculationOperation - start porting optimized Jack generator to fermions
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1419 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ src/Architecture/ArchitectureOperation/AbstractPrecalculationOperation.h
@@ -46,9 +46,9 @@ class AbstractPrecalculationOperation: public AbstractArchitectureOperation
   // number of component 
   int NbrComponent;
   // index of the first component (if long numbers are required)
-  int LargeFirstComponent;
+  long LargeFirstComponent;
   // number of component  (if long numbers are required)
-  int LargeNbrComponent;
+  long LargeNbrComponent;
 
  public:
